Fix some bugs in Hydra.
Exit-printing is still too fiddly for my liking.

Woods' Adventure gives a suboptimal response to the chain of
commands "LAMP" (What do you want to do with the lamp?) "BLAH"
("Sorry, I don't know the word blah") "QUIT" (Eh?) because it
is still listening for what to do with the lamp and therefore
believes that last input means "QUIT LAMP". I think an unrecognized
word should reset the parser; so, done.